WIP: use new confirm menu
