+++ src/actions.js
@@ -3,7 +3,6 @@ export {
   cancelMoves,
   confirmMoves,
   makeMove,
-  postMove,
   selectMove,
 } from 'scenes/Game/actions';
