+++ src/scenes/Game/components/ConfirmMoveMenu.js
@@ -1,7 +1,7 @@
 import React, { useCallback } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 
-import { cancelMoves, confirmMoves, postMove } from 'actions';
+import createMoveOnServer from 'actions/createMoveOnServer';
 import { getLastMove } from 'reducers';
 
 import isEqual from 'lodash/isEqual';
@@ -12,30 +12,32 @@ const ConfirmMoveMenu = () => {
   const dispatch = useDispatch();
 
   const gameSlug = useSelector(state => state.gameSlug);
+  const showConfirmMoveMenu = useSelector(
+    state => state.showConfirmMoveMenu,
+  );
   const lastMove = useSelector(state => getLastMove(state), isEqual);
   const username = useSelector(state => state.username);
 
-  const confirmGameMove = useCallback(() => {
-    dispatch(postMove({ gameSlug, move: lastMove, username }));
-    dispatch(confirmMoves());
-  }, [dispatch, gameSlug, lastMove, username]);
+  const confirmPosition = useCallback(
+    async () => {
+      dispatch(createMoveOnServer({ gameSlug, position: lastMove, username }));
+      dispatch({ type: 'toggle_show_confirm_move_menu', value: false });
+    },
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+    [dispatch, gameSlug, lastMove.id, username],
+  );
 
-  const confirmMove = useCallback(async () => {
-    if (!lastMove.pending) return;
-
-    confirmGameMove();
-  }, [confirmGameMove, lastMove.pending]);
-
-  const cancelMove = useCallback(() => {
-    dispatch(cancelMoves());
-  }, [dispatch]);
+  const cancelPosition = useCallback(() => {
+    dispatch({ type: 'toggle_show_confirm_move_menu', value: false });
+    dispatch({ type: 'remove_position', id: lastMove.id });
+  }, [dispatch, lastMove.id]);
 
   return (
     <ConfirmMenu
       label="Confirm move?"
-      yesHandler={confirmMove}
-      noHandler={cancelMove}
-      show={lastMove.pending}
+      yesHandler={confirmPosition}
+      noHandler={cancelPosition}
+      show={showConfirmMoveMenu}
       disabled={gameSlug === null}
     />
   );
